Rename movie container
